-aggiunta l'esplosione.

+++ cube.h
@@ -88,10 +88,10 @@ private:
     GLint levelCellsWidth; /**< It is the number of cells of the Level's width. */
     GLfloat scaleFactor; /**< It is the scale factor. */
     GLfloat jumpStartTime; /**< It is the time the Cube has started the jump. */
-    GLuint jumpParameter; /**< TODO */
-    GLfloat jumpVelocity;  /**< TODO */
-    GLfloat movingParameter;  /**< TODO */
-    GLuint t;  /**< It is the parametric variable to put in the parametric functions. */
+    GLuint jumpParameter; /**< It is the parametric variable to put in the jump parametric functions. */
+    GLfloat jumpVelocity;  /**< It is the velocity of the jump. */
+    GLuint movingParameter;  /**< It is the parametric variable to put in the moving parametric functions */
+    GLuint t; /**< TODO */
     Vector3f* normalsMatrix[4][4][4]; /**< It is the 3-dimensional matrix that contain the nomrs vector of each sub-cube in case of explosion. */
 
     /**
